# api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/dedup/scan": {
      "post": {
        "operationId": "DeduplicationController_scan",
        "parameters": [
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerScanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Deduplication disabled or job already running"
          }
        },
        "summary": "Trigger batch deduplication scan",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/scan/{scanId}": {
      "get": {
        "operationId": "DeduplicationController_getScanStatus",
        "parameters": [
          {
            "name": "scanId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Scan not found"
          }
        },
        "summary": "Get batch scan status",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/candidates": {
      "get": {
        "operationId": "DeduplicationController_getCandidates",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SKIPPED"
              ]
            }
          },
          {
            "name": "minSimilarity",
            "required": false,
            "in": "query",
            "description": "Minimum similarity filter",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum results to return",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCandidatesResponseDto"
                }
              }
            }
          }
        },
        "summary": "List merge candidates",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/candidates/{candidateId}": {
      "get": {
        "operationId": "DeduplicationController_getCandidate",
        "parameters": [
          {
            "name": "candidateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Candidate not found"
          }
        },
        "summary": "Get merge candidate by ID",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/review/{candidateId}/approve": {
      "post": {
        "operationId": "DeduplicationController_approve",
        "parameters": [
          {
            "name": "candidateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-approver-id",
            "required": false,
            "in": "header",
            "description": "Approver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Candidate not found"
          }
        },
        "summary": "Approve merge candidate",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/review/{candidateId}/reject": {
      "post": {
        "operationId": "DeduplicationController_reject",
        "parameters": [
          {
            "name": "candidateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-approver-id",
            "required": false,
            "in": "header",
            "description": "Rejector ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Candidate not found"
          }
        },
        "summary": "Reject merge candidate",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/review/{candidateId}/skip": {
      "post": {
        "operationId": "DeduplicationController_skip",
        "parameters": [
          {
            "name": "candidateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "Days to skip (default 7)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Candidate not found"
          }
        },
        "summary": "Skip merge candidate",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/candidates/bulk-resolve": {
      "post": {
        "operationId": "DeduplicationController_bulkResolve",
        "parameters": [
          {
            "name": "x-approver-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Bulk resolve merge candidates",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/candidates/expire-stale": {
      "post": {
        "operationId": "DeduplicationController_expireStale",
        "parameters": [
          {
            "name": "ttlDays",
            "required": false,
            "in": "query",
            "description": "TTL in days (default 14)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Auto-resolve PENDING candidates older than TTL",
        "tags": [
          "deduplication"
        ]
      }
    },
...
    "/v1/dedup/merge": {
      "post": {
        "operationId": "DeduplicationController_merge",
        "parameters": [
          {
            "name": "x-approver-id",
            "required": false,
            "in": "header",
            "description": "Approver ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualMergeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "summary": "Manually merge memories",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/merge/{mergeEventId}/rollback": {
      "post": {
        "operationId": "DeduplicationController_rollback",
        "parameters": [
          {
            "name": "mergeEventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot rollback"
          },
          "404": {
            "description": "Merge event not found"
          }
        },
        "summary": "Rollback a merge",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/history": {
      "get": {
        "operationId": "DeduplicationController_getHistory",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "survivorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get merge history",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/history/{mergeEventId}": {
      "get": {
        "operationId": "DeduplicationController_getMergeEvent",
        "parameters": [
          {
            "name": "mergeEventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeEventDto"
                }
              }
            }
          },
          "404": {
            "description": "Merge event not found"
          }
        },
        "summary": "Get merge event details",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/lineage/{memoryId}": {
      "get": {
        "operationId": "DeduplicationController_getLineage",
        "parameters": [
          {
            "name": "memoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get memory lineage",
        "tags": [
          "deduplication"
        ]
      }
    },
    "/v1/dedup/similar/{memoryId}": {
      "get": {
        "operationId": "DeduplicationController_findSimilar",
        "parameters": [
          {
            "name": "memoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topK",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minSimilarity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimilarMemoryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Find similar memories",
        "tags": [
          "deduplication"
        ]
      }
    },
...
    "/v1/dedup/stats": {
      "get": {
        "operationId": "DeduplicationController_getStats",
        "parameters": [
          {
            "name": "x-am-user-id",
            "in": "header",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get deduplication statistics",
        "tags": [
          "deduplication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Engram API",
    "description": "Memory infrastructure for AI agents",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-API-Key"
      },
      "user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-User-ID"
      }
    },
    "schemas": {
...
      "TriggerScanDto": {
        "type": "object",
        "properties": {
          "dryRun": {
            "type": "boolean",
            "description": "Run in dry-run mode (no actual merges)"
          },
          "minSimilarity": {
            "type": "number",
            "description": "Minimum similarity threshold (0.0-1.0)"
          },
          "maxMemories": {
            "type": "number",
            "description": "Maximum memories to process"
          },
          "userId": {
            "type": "string",
            "description": "Filter to specific user ID"
          }
        }
      },
      "ScanResponseDto": {
        "type": "object",
        "properties": {
          "scanId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "memoriesProcessed": {
            "type": "number"
          },
          "clustersFound": {
            "type": "number"
          },
          "autoMerged": {
            "type": "number"
          },
          "queuedForReview": {
            "type": "number"
          },
          "durationMs": {
            "type": "number"
          }
        },
        "required": [
          "scanId",
          "status",
          "memoriesProcessed",
          "clustersFound",
          "autoMerged",
          "queuedForReview",
          "durationMs"
        ]
      },
...
      "ListCandidatesResponseDto": {
        "type": "object",
        "properties": {
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MergeCandidateDto"
            }
          },
          "total": {
            "type": "number"
          },
          "pendingCount": {
            "type": "number"
          }
        },
        "required": [
          "candidates",
          "total",
          "pendingCount"
        ]
      },
...
    }
  }
}